CLA: trivial
Fix double free crash.  This occurs when using a provider property string.  For example, passing in "fips=no" to SSL_CTX_new_ex causes the internal PROV_MAC_CTX struct to set the propq field to that value.  This structure is duplicated and freed in calls to SSL_write but the duplicate method (mac_dupctx) does a shallow copy which is then freed.  The next call to SSL_write or SSL_free will cause a double free.